# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/countries": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Get a listing of the Countries.",
                "description": "Get all Countries",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Country"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Country"
                ],
                "summary": "Store a newly created Country in storage",
                "description": "Store Country",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Country that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Country"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Country"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/countries/{id}": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Display the specified Country",
                "description": "Get Country",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Country",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Country"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Country"
                ],
                "summary": "Update the specified Country in storage",
                "description": "Update Country",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Country",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Country that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Country"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Country"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Country"
                ],
                "summary": "Remove the specified Country from storage",
                "description": "Delete Country",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Country",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Country": {
            "required": [
                "capital",
                "citizenship",
                "country_code",
                "currency",
                "currency_code",
                "currency_sub_unit",
                "currency_symbol",
                "full_name",
                "iso_3166_2",
                "iso_3166_3",
                "name",
                "region_code",
                "sub_region_code",
                "eea",
                "calling_code",
                "flag"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "capital": {
                    "description": "capital",
                    "type": "string"
                },
                "citizenship": {
                    "description": "citizenship",
                    "type": "string"
                },
                "country_code": {
                    "description": "country_code",
                    "type": "string"
                },
                "currency": {
                    "description": "currency",
                    "type": "string"
                },
                "currency_code": {
                    "description": "currency_code",
                    "type": "string"
                },
                "currency_sub_unit": {
                    "description": "currency_sub_unit",
                    "type": "string"
                },
                "currency_symbol": {
                    "description": "currency_symbol",
                    "type": "string"
                },
                "full_name": {
                    "description": "full_name",
                    "type": "string"
                },
                "iso_3166_2": {
                    "description": "iso_3166_2",
                    "type": "string"
                },
                "iso_3166_3": {
                    "description": "iso_3166_3",
                    "type": "string"
                },
                "name": {
                    "description": "name",
                    "type": "string"
                },
                "region_code": {
                    "description": "region_code",
                    "type": "string"
                },
                "sub_region_code": {
                    "description": "sub_region_code",
                    "type": "string"
                },
                "eea": {
                    "description": "eea",
                    "type": "boolean"
                },
                "calling_code": {
                    "description": "calling_code",
                    "type": "string"
                },
                "flag": {
                    "description": "flag",
                    "type": "string"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        }
    }
}